--- src/model/inners.js
@@ -1,5 +1,0 @@
-import { Node } from './node'
-
-export const Catch = (variable, type) => (...handler) => Node('Catch')({ variable, type, handler: Block(...handler) })
-export const Parameter = (name, varArg = false) => Node('Param')({ name, varArg })
-export const Block = (...sentences) => Node('Block')({ sentences })
